moving a service into the correct directory
