+++ web/src/app/pages/profile/update-password/update-password.component.ts
@@ -10,7 +10,7 @@ import { HeroComponent } from 'src/app/shared/hero/hero.component';
 import { FjorgeUser, UserModel } from 'src/app/shared/interfaces/user.model';
 import { IconComponent } from 'src/app/shared/icon/icon.component';
 import { ToastComponent } from 'src/app/shared/toast/toast.component';
-import { ToastService } from 'src/app/services/toast.service';
+import { ToastService } from 'src/services/toasts/toast.service';
 
 @Component({
   selector: 'app-update-password',
