+++ web/src/app/shared/form-update-facility/form-update-facility.component.ts
@@ -7,7 +7,7 @@ import { AccordionComponent } from '../accordion/accordion.component';
 import { ButtonComponent } from '../button/button.component';
 import { IconComponent } from '../icon/icon.component';
 import { ToastComponent } from '../toast/toast.component';
-import { ToastService } from 'src/app/services/toast.service';
+import { ToastService } from 'src/services/toasts/toast.service';
 import { Router } from '@angular/router';
 
 @Component({
